requirements.txt
Django==1.6.1
South==0.8.4
django-appconf==0.6
django-bower==4.8.1
django-compressor==1.3
django-jstemplate==1.3.6
django-registration==1.0
djangorestframework==2.3.10
ipython==1.1.0
mwparserfromhell==0.3.2
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.1.0
six==1.4.1
wsgiref==0.1.2
